I realized that with the partial match way that duris works now, if a title bar has 'god gods' on it, it's redundant. if I do help "god" i'll be found if the title bar is just 'gods' because of the partial match workup.
I'm going to go through over the next few days and set everything
to plural, so that when it is singular or plural that is searched,
it'll show up for both, and reduce the length of all the helpfile
titles, as well as filesize.

This is a rather minor thing for the current help system, but it does
help a great deal for my future system.