Drop dependency between age and hp/mv/mana max and regen.
It doesn't add anything to the game but penalize people who happen to
forget to ask a necro to cast a spell a few times once.

The common wisdom that "it's best to go with 2× the starting age" is
incorrect -- eg. being older than that gives you more moves but less
move regen, etc.  But, it gives us a nice fixed point.

No cleanup of dead code to reduce patch conflicts.